+++ Cargo.toml
@@ -166,7 +166,6 @@ wasmtime-winch = { path = "crates/winch", version = "=20.0.0" }
 wasmtime-environ = { path = "crates/environ", version = "=20.0.0" }
 wasmtime-explorer = { path = "crates/explorer", version = "=20.0.0" }
 wasmtime-fiber = { path = "crates/fiber", version = "=20.0.0" }
-wasmtime-fibre = { path = "crates/fibre", version = "=20.0.0" }
 wasmtime-types = { path = "crates/types", version = "20.0.0" }
 wasmtime-jit-debug = { path = "crates/jit-debug", version = "=20.0.0" }
 wasmtime-runtime = { path = "crates/runtime", version = "=20.0.0" }
