+++ crates/continuations/Cargo.toml
@@ -14,5 +14,4 @@ repository = "https://github.com/wasmfx/wasmtime"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-wasmtime-fibre = { workspace = true }
 memoffset = "0.9.0"
