--- crates/fibre/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "wasmtime-fibre"
-version.workspace = true
-authors.workspace = true
-description = "Fiber support for Wasmtime"
-license = "Apache-2.0 WITH LLVM-exception"
-repository = "https://github.com/bytecodealliance/wasmtime"
-edition.workspace = true
-
-# We link to some native code with symbols that don't change often, so let Cargo
-# know that we can't show up multiple times in a crate graph. If this is an
-# issue in the future we should tweak the build script to set `#define`
-# directives or similar to embed a version number of this crate in symbols.
-links = "wasmtime-fibre-shims"
-
-[dependencies]
-cfg-if = { workspace = true }
-
-[target.'cfg(unix)'.dependencies]
-rustix = { workspace = true, features = ["mm", "param"] }
-wasmtime-asm-macros = { workspace = true }
-
-[build-dependencies]
-cc = "1.0"
-
-[dev-dependencies]
-backtrace = "0.3.61"
