+++ crates/runtime/src/lib.rs
@@ -30,6 +30,7 @@ mod table;
 mod traphandlers;
 mod vmcontext;
 
+pub mod fibre;
 pub mod continuation;
 #[cfg(feature = "debug-builtins")]
 pub mod debug_builtins;
